Finished basic lexer impl
Still one FIXME that is TBD:

The current lexer records `10` as `[Token::Int(1), Token::Int(0)]`. I would like it or expected it to record `10` as `[Token::Int(10)]`.
This is TBD as I'm not sure how to impl this and I'm not sure if it's nessary. We could make ajustments in the AST or interpreter later on which
doesn't require `10` to be recorded as `[Token::Int(10)]` but I'm not sure if that would work.

TBD

Other than that, the current lexer impl works as expected.